requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
aiosqlite==0.18.0
async-timeout==4.0.2
attrs==22.2.0
auraxium==0.2.3
backoff==2.2.1
charset-normalizer==3.1.0
discord.py==2.2.2
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
pydantic==1.10.5
python-dotenv==1.0.0
typing_extensions==4.5.0
websockets==10.4
yarl==1.8.2
redis==4.5.4